src/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
imageio==2.36.1
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.1
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-image==0.25.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.5.0
tifffile==2024.12.12
tzdata==2024.2
